+++ examples/06_error_handling.rs
@@ -28,10 +28,10 @@ fn main() {
     // You can optionally define a custom error handler
     // This can be defined anywhere in the server and will take affect for all routes
     // Its like a normal route, but it will only be called if the route panics
-    server.set_error_handler(|_req| {
+    server.set_error_handler(|_req, err| {
         Response::new(
             500,
-            "Internal Server Error: Something bad happened",
+            &format!("<h1>Internal Server Error</h1><br>Panicked at '{}'", err),
             vec![Header::new("Content-Type", "text/html")],
         )
     });