+++ lib/lib.rs
@@ -81,9 +81,10 @@ server.start_threaded(8);
 
 #![warn(missing_docs)]
 
-pub(crate) const VERSION: &str = "0.1.6";
+pub(crate) const VERSION: &str = "0.1.7*";
 
 mod common;
+mod http;
 #[cfg(feature = "thread_pool")]
 mod threadpool;